[S09P22B310-000] feat : makeTransaction추가

+++ backend/src/main/java/com/ntt/mwonimoney/domain/account/api/v1/FinAccountApi.java
@@ -395,7 +395,7 @@ public class FinAccountApi {
 
 	@GetMapping("/accounts/small-account/{finAccountIdx}/remain")
 	public ResponseEntity getRemain(@PathVariable Long finAccountIdx) {
-		Long remain = finAccountService.getBalance(finAccountIdx);
+		int remain = finAccountService.getBalance(finAccountIdx);
 		return ResponseEntity.ok().body(remain);
 	}
